Chat: prevent unnecessary message wipe

+++ examples/flux-chat/js/ChatExampleData.js
@@ -13,6 +13,7 @@
 module.exports = {
 
   init: function() {
+    if(localStorage.hasOwnProperty('messages')) return;
     localStorage.clear();
     localStorage.setItem('messages', JSON.stringify([
       {
